Set projects to use Swift 5. Version bump. Update Package.swift. Use legacy build system (for Carthage compatibility). Separate extensions for CMNode.
