+++ Maaku.podspec
@@ -1,7 +1,7 @@
 Pod::Spec.new do |s|
 
   s.name         = "Maaku"
-  s.version      = "0.8.0"
+  s.version      = "0.8.1"
   s.summary      = "Swift cmark-gfm wrapper with a Swift friendly representation of the AST"
 
   s.description  = <<-DESC
@@ -16,6 +16,7 @@ Pod::Spec.new do |s|
   s.osx.deployment_target = "10.11"
   s.tvos.deployment_target = "9.0"
   s.watchos.deployment_target = "2.0"
+  s.swift_version = "5.0"
 
   s.source       = { :git => "https://github.com/KristopherGBaker/Maaku.git", :tag => "#{s.version}" }
   s.requires_arc = true
